Retarget URLs using a click handler
Unlike the previous method, this lets me keep the default href - and
therefore mouseover - pointing to a real URL that does trigger a full
page reload, but because we event.preventDefault() in our click handler,
the browser doesn't actually follow the href. This is what I wanted from
the beginning, but a problem where I wasn't calling retargetMungeLinks()
at the right time caused me to believe this approach was not workable.
After fixing that two commits ago, this method works.