Add ro and rw LockableSqliteConnection properties
Change the way that LockableSqliteConnection works so that it is not
itself a with-able object, but instead contains two with-able
properties, .ro and .rw. The .ro property is intended for *non*locking
read operations, while the .rw property is intended for locking write
operations. Allows us to reduce database locking.