BROKEN: Experiments w/ AppEndpoints
I did some experimentation with an structre I called AppEndpoints. They
were custom objects that knew how to call backend code, named after the
backend function. I still think the idea is interesting, but I think
they're too heavyweight for this project, so I'm recording them (in a
broken state!) in this commit, and I will revert it in but a moment.