Do not overload the application() function
Previously, application() was called when running under the CherryPy
webserver, and it's also part of the WSGI API. This never actually
worked, although it's been in the codebase for years at this point.

Break out the application() function so that it's generated by CherryPy,
and use a totally separate function for non-WSGI stuff.